add more safety belts

+++ main.go
@@ -75,8 +75,30 @@ func walkDir(path string, d fs.DirEntry, err error) (e error) {
 }
 
 /*=================
-  Copy functions
+  File functions
 */
+
+type RemoveFunc func(src string) error
+
+func removeDry(src string) error {
+	fmt.Printf("remove %s\n", src)
+	return nil
+}
+
+func removeFile(src string) error {
+	return os.Remove(src)
+}
+
+type RenameFunc func(src, dst string) error
+
+func renameDry(src, dst string) error {
+	fmt.Printf("rename %s -> %s\n", src, dst)
+	return nil
+}
+
+func renameFile(src, dst string) error {
+	return os.Rename(src, dst)
+}
 
 type CopyFunc func(src, dst string) (int64, error)
 
@@ -108,6 +130,10 @@ func copyFile(src, dst string) (int64, error) {
 		return 0, err
 	}
 
+	exists := fileExists(dst)
+	if exists != 0 {
+		return 0, errors.New(dst + " exists")
+	}
 	var destination *os.File
 	destination, err = os.Create(dst)
 	if err != nil {
@@ -136,10 +162,12 @@ func fileExists(dst string) int {
 
 func install(dryRun, skipErrors bool) {
 	var (
-		basePath string
-		checks   bool
-		copy     CopyFunc
-		exists   int
+		basePath      string
+		checks        bool
+		copy          CopyFunc
+		rename        RenameFunc
+		exists        int
+		criticalError bool
 	)
 
 	if tgtDir != "" {
@@ -149,59 +177,81 @@ func install(dryRun, skipErrors bool) {
 	}
 	filepath.WalkDir(moduleRootPath, walkDir)
 	// filepath.Walk(filepath.Join(nwnDir, extensionsDir, moduleName), walker)
-	fmt.Println("Run pre-install checks...")
+	fmt.Println("> Run pre-install checks...")
 
 	checks = true
 
 	if dryRun {
 		copy = copyDry
+		rename = renameDry
 	} else {
 		copy = copyFile
+		rename = renameFile
 	}
 
+	criticalError = false
 	for _, fpath := range moduleInfo.Files {
 		dst := filepath.Join(basePath, fpath)
 		exists = fileExists(dst)
 		excluded := moduleInfo.Excluded.Contains(fpath)
 		if exists != 0 {
 			// file exists
-			fmt.Printf("%s seems to exist\n", fpath)
+			// fmt.Printf("%s seems to exist\n", fpath)
 			checks = false
 		} else if !excluded {
 			// filed does not exist and is not excluded
 			moduleInfo.Installed = append(moduleInfo.Installed, fpath)
 		}
 		if exists != 0 && moduleInfo.OverwriteExisting && !excluded {
-			// copy(dst, dst+moduleInfo.BackupExtension)
-			moduleInfo.Saved = append(moduleInfo.Saved, fpath+moduleInfo.BackupExtension)
-			moduleInfo.Installed = append(moduleInfo.Installed, fpath)
+			if fileExists(dst+moduleInfo.BackupExtension) == 0 {
+				// Backup does not exist
+				moduleInfo.Saved = append(moduleInfo.Saved, fpath+moduleInfo.BackupExtension)
+				moduleInfo.Installed = append(moduleInfo.Installed, fpath)
+			} else {
+				// Backup exists already. Danger zone
+				fmt.Printf("! %s already exist\n", fpath+moduleInfo.BackupExtension)
+				criticalError = true
+			}
 		} else if exists != 0 && !excluded {
+			fmt.Printf("! %s skipped\n", fpath)
 			moduleInfo.Skipped = append(moduleInfo.Skipped, fpath)
 		}
 	}
 
-	if !checks && !skipErrors && !moduleInfo.OverwriteExisting {
-		fmt.Printf("Exiting due to previous errors\n")
+	if (!checks && !skipErrors && !moduleInfo.OverwriteExisting) || criticalError {
+		fmt.Printf("! Exiting due to previous errors\n")
 		os.Exit(1)
 	}
 
 	for _, fpath := range moduleInfo.Installed {
 		// fmt.Println(fpath)
+		operation := "+"
 		if moduleInfo.Saved.Contains(fpath + moduleInfo.BackupExtension) {
+			// Save backup copy
 			dst := filepath.Join(basePath, fpath)
-			fmt.Printf("%s is being backed up\n", fpath)
-			copy(dst, dst+moduleInfo.BackupExtension)
+			operation = "^"
+			// copy(dst, dst+moduleInfo.BackupExtension)
+			if fileExists(dst+moduleInfo.BackupExtension) == 0 {
+				rename(dst, dst+moduleInfo.BackupExtension)
+			}
 		}
+		fmt.Printf("%s %s\n", operation, fpath)
 		copy(filepath.Join(moduleRootPath, fpath), filepath.Join(basePath, fpath))
 	}
 	manifest, _ := json.Marshal(moduleInfo)
 	// fmt.Println(string(manifest))
 	manifestFile := filepath.Join(moduleInfo.NwnDir, moduleInfo.ExtensionsDir, moduleInfo.Name+".json")
-	fmt.Printf("Writing manifest to %s\n", manifestFile)
-	ioutil.WriteFile(manifestFile, manifest, 0640)
+	fmt.Printf("= Writing manifest to %s\n", manifestFile)
+	if !dryRun {
+		ioutil.WriteFile(manifestFile, manifest, 0640)
+	}
 }
 
 func uninstall(dryRun, skipErrors bool) {
+	var (
+		rename RenameFunc
+		remove RemoveFunc
+	)
 	manifestFile := filepath.Join(moduleInfo.NwnDir, moduleInfo.ExtensionsDir, moduleInfo.Name+".json")
 	uninstalledFile := filepath.Join(moduleInfo.NwnDir, moduleInfo.ExtensionsDir, moduleInfo.Name+".uninstalled")
 	jsonFile, err := os.Open(manifestFile)
@@ -209,6 +259,13 @@ func uninstall(dryRun, skipErrors bool) {
 		fmt.Println(err)
 	}
 	defer jsonFile.Close()
+	if dryRun {
+		rename = renameDry
+		remove = removeDry
+	} else {
+		rename = renameFile
+		remove = removeFile
+	}
 
 	byteValue, _ := ioutil.ReadAll(jsonFile)
 	var mi ModuleInfo
@@ -216,33 +273,32 @@ func uninstall(dryRun, skipErrors bool) {
 	if err != nil {
 		panic(err)
 	}
+
+	//XXX Override base vars with the one from CLI (?)
+	mi.NwnDir = moduleInfo.NwnDir
+	mi.ExtensionsDir = moduleInfo.ExtensionsDir
+	mi.Name = moduleInfo.Name
+
 	for _, installed := range mi.Installed {
 		filepath := filepath.Join(mi.NwnDir, installed)
-		if !dryRun {
-			os.Remove(filepath)
-		} else {
-			fmt.Printf("- %s\n", filepath)
-		}
+		fmt.Printf("- %s\n", filepath)
+		remove(filepath)
 	}
 	for _, saved := range mi.Saved {
 		filepath := filepath.Join(mi.NwnDir, saved)
 		// we need to rename filepath to original by removing suffix
 		// strings.HasSuffix(xxx,mi.BackupExtension)
 		originalFilepath := strings.TrimSuffix(filepath, mi.BackupExtension)
-		if !dryRun {
-			err := os.Rename(filepath, originalFilepath)
-			if err != nil {
-				fmt.Println(err)
-			}
-		} else {
-			fmt.Printf("o %s <- %s\n", originalFilepath, filepath)
-		}
-	}
-	if !dryRun {
-		err = os.Rename(manifestFile, uninstalledFile)
+		err := rename(filepath, originalFilepath)
 		if err != nil {
 			fmt.Println(err)
 		}
+		fmt.Printf("v %s\n", filepath)
+	}
+
+	err = rename(manifestFile, uninstalledFile)
+	if err != nil {
+		fmt.Println(err)
 	}
 }
